Add "Constant ID" to Trainer class
Adds an int value to the Trainer class, "constantId". This value is used as the ID for the constants file when outputting to the game, so that trainer IDs do not change in between uses (which would break save files in Fandango). When creating a new trainer, its constant ID is assigned when saving and appended to the end of the list.